Add new test for pkg explorer diff pkg view * Remove unused function from pkg explorer * Try to fix windows path in coverage * Split gui_etc tests in 2 files * overhaul some testcases
